Module Bindlib.Lift

Functorial interface used to build lifting functions for any type equipped with a map function. In other words, this function can be used to allow the permutation of the 'a box type with another type constructor.

Parameters

M : Map

Signature

val lift_box : 'a box M.t -> 'a M.t box